# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Dokumentasi interaktif RESTful API Pengelolaan User, Kategori, dan Buku beserta Upload Cover Gambar.",
        "title": "Book \u0026 Category Management REST API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Harits Nala Barrun",
            "email": "developer.haritsnb@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/users/login": {
            "post": {
                "description": "Otentikasi kredensial pengguna untuk mendapatkan JSON Web Token (JWT)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "User Login",
                "parameters": [
                    {
                        "description": "Payload Login",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UserLoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login berhasil dan mengembalikan token",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Format input tidak valid",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Username atau password salah",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/users/logout": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Memasukkan token JWT yang sedang aktif ke dalam Blacklist memori",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "User Logout",
                "responses": {
                    "200": {
                        "description": "Logout berhasil",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized / Token tidak valid",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.UserLoginRequest": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Masukkan token dengan format: Bearer \u003cTOKEN_JWT_ANDA\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}